+++ CHANGELOG.md
@@ -17,6 +17,7 @@ User interface overhaul.
 - Tweaked the SplashScreen state animation.
 - SplashScreen and MainMenu states now use the Mars sound track in the background.
 - Setting minimum resolution to 1000x700. I know, kind of a weird one but this is intended mostly for windowed modes to run properly on low-resolution displays (1366x768).
+- Structures that are disabled now indicate why they were disabled.
 
 ### Fixed
 - Fixed an issue that prevented the planet description from displaying.